requirements.txt
streamlit==1.28.1
google-generativeai==0.3.0
python-docx==0.8.11
flask==3.0.0
